Divide classes into packages blue, blue.task and blue.handler
