--- src/main/java/BlueException.java
@@ -1,5 +1,0 @@
-class BlueException extends Exception {
-    BlueException(String message) {
-        super(message);
-    }
-}
