+++ src/main/java/blue/BlueException.java
@@ -1,0 +1,7 @@
+package blue;
+
+public class BlueException extends Exception {
+    public BlueException(String message) {
+        super(message);
+    }
+}
