+++ src/main/java/blue/TaskList.java
@@ -1,3 +1,7 @@
+package blue;
+
+import blue.task.Task;
+
 import java.util.ArrayList;
 import java.util.List;
