+++ src/main/java/blue/handler/ListHandler.java
@@ -1,10 +1,14 @@
+package blue.handler;
+
+import blue.TaskList;
+
 public class ListHandler extends CommandHandler {
     public ListHandler(TaskList taskList) {
         super(taskList);
     }
 
     @Override
-    String handle(String input) {
+    public String handle(String input) {
         String response = "Here are the tasks in your list:\n";
         String[] lines = new String[taskList.size()];
         for (int i = 1; i <= taskList.size(); i++) {
