+++ src/main/java/blue/task/Task.java
@@ -1,4 +1,6 @@
-abstract class Task {
+package blue.task;
+
+public abstract class Task {
     private final String title;
     private boolean isDone;
